+++ randovania/games/am2r/gui/preset_settings/__init__.py
@@ -10,6 +10,7 @@ if TYPE_CHECKING:
 
 
 def preset_tabs(editor: PresetEditor, window_manager: WindowManager):
+    from randovania.games.am2r.gui.preset_settings.am2r_chaos_tab import PresetAM2RChaos
     from randovania.games.am2r.gui.preset_settings.am2r_goal_tab import PresetAM2RGoal
     from randovania.games.am2r.gui.preset_settings.am2r_hints_tab import PresetAM2RHints
     from randovania.games.am2r.gui.preset_settings.am2r_patches_tab import PresetAM2RPatches
@@ -33,4 +34,5 @@ def preset_tabs(editor: PresetEditor, window_manager: WindowManager):
         PresetDockRando,
         PresetTeleportersAM2R,
         PresetAM2RPatches,
+        PresetAM2RChaos,
     ]